+++ packages/web-forms/src/index.ts
@@ -1,10 +1,17 @@
 import { webFormsPlugin } from './WebFormsPlugin';
 import OdkWebForm from './components/OdkWebForm.vue';
 
-import './assets/css/icomoon.css';
-import './themes/2024-light/theme.scss';
+import icomoon from './assets/css/icomoon.css?inline';
+import theme from './themes/2024-light/theme.scss?inline';
 
 // TODO/sk: Purge it - using postcss-purgecss
-import 'primeflex/primeflex.css';
+import primeflex from 'primeflex/primeflex.css?inline';
+
+const styles = [icomoon, theme, primeflex].join('\n\n');
+const stylesheet = new CSSStyleSheet();
+
+stylesheet.replaceSync(styles);
+
+document.adoptedStyleSheets.push(stylesheet);
 
 export { OdkWebForm, webFormsPlugin };