+++ packages/common/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @getodk/xforms-engine
+# @getodk/common
 
 ## 0.1.1