+++ packages/tree-sitter-xpath/CHANGELOG.md
@@ -1,1 +1,8 @@
 # @getodk/tree-sitter-xpath
+
+## 0.1.1
+
+### Patch Changes
+
+- aa0e214: Update to latest stable Node versions: 18.20.3, 20.13.1, 22.2.0
+- a26788b: Updated external dependencies