+++ packages/ui-solid/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @getodk/xforms-engine
+# @getodk/ui-solid
 
 ## 0.1.1
 
@@ -6,3 +6,7 @@
 
 - aa0e214: Update to latest stable Node versions: 18.20.3, 20.13.1, 22.2.0
 - a26788b: Updated external dependencies
+- Updated dependencies [aa0e214]
+- Updated dependencies [a26788b]
+  - @getodk/xforms-engine@0.1.1
+  - @getodk/common@0.1.1