+++ packages/xpath/CHANGELOG.md
@@ -1,1 +1,11 @@
 # @getodk/xpath
+
+## 0.1.1
+
+### Patch Changes
+
+- aa0e214: Update to latest stable Node versions: 18.20.3, 20.13.1, 22.2.0
+- a26788b: Updated external dependencies
+- Updated dependencies [aa0e214]
+- Updated dependencies [a26788b]
+  - @getodk/common@0.1.1